feat(layout): Header
Created a layout folder to contain all layout components, created a header folder inside the layout folder to contain the header component and all its possible variants and css module files.

Created a Header component in the Header folder and created a Header.module.css file for header styles.

Created a types folder in the app folder to contain all types for the app, inside the types folder, created a layout.d.ts file for layout types and updated the file with header types.

In the folder export, created a styles.ts file to export all css modules.

In the file tsconfig.json, included the path to the styles export file.